Move settings to specific module package
