+++ settings/.metadata
@@ -1,0 +1,10 @@
+# This file tracks properties of this Flutter project.
+# Used by Flutter tool to assess capabilities and perform upgrades etc.
+#
+# This file should be version controlled and should not be manually edited.
+
+version:
+  revision: fb57da5f945d02ef4f98dfd9409a72b7cce74268
+  channel: stable
+
+project_type: package
