+++ settings/lib/settings/settings_page.dart
@@ -1,9 +1,12 @@
-import 'package:tankste/version_item.dart';
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import 'package:settings/version/ui/version_item.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 class SettingsPage extends StatelessWidget {
+
+  const SettingsPage({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     final List<Widget> items = _buildItems(context);
